LayoutTests/inspector/indexeddb/requestDatabaseNames-expected.txt

CONSOLE MESSAGE: line 10: Created Database 'Database1'
CONSOLE MESSAGE: line 10: Created Database 'Database2'
CONSOLE MESSAGE: line 10: Created Database 'á‰'
CONSOLE MESSAGE: line 10: Created Database 'ğ†'

PASS: No IndexedDB databases should exist initially
Created Database 'Database1'
PASS: A single IndexedDB database should exist
["Database1"]
Created Database 'Database2'
PASS: Two IndexedDB databases should exist
["Database1","Database2"]
Created Database 'á‰'
PASS: Two IndexedDB databases should exist
["Database1","Database2","á‰"]
Created Database 'ğ†'
PASS: Four IndexedDB databases should exist
["Database1","Database2","á‰","ğ†"]
